+++ Controller/Admin/Post.php
@@ -117,6 +117,7 @@ final class Post extends AbstractAdminController
      */
     public function deleteAction($id)
     {
+        $historyService = $this->getService('Cms', 'historyManager');
         $service = $this->getModuleService('postManager');
 
         // Batch removal
@@ -126,14 +127,21 @@ final class Post extends AbstractAdminController
             $service->deleteByIds($ids);
             $this->flashBag->set('success', 'Selected elements have been removed successfully');
 
+            // Save in the history
+            $historyService->write('News', '%s posts have been removed', count($ids));
+
         } else {
             $this->flashBag->set('warning', 'You should select at least one element to remove');
         }
 
         // Single removal
         if (!empty($id)) {
+            $post = $this->getPostManager()->fetchById($id, false, false);
             $service->deleteById($id);
             $this->flashBag->set('success', 'Selected element has been removed successfully');
+
+            // Save in the history
+            $historyService->write('News', 'Post "%s" has been removed', $post->getName());
         }
 
         return '1';
@@ -187,15 +195,22 @@ final class Post extends AbstractAdminController
         if (1) {
             $service = $this->getModuleService('postManager');
 
+            $historyService = $this->getService('Cms', 'historyManager');
+            $name = $this->getCurrentProperty($this->request->getPost('translation'), 'name');
+
             if (!empty($post['id'])) {
                 if ($service->update($this->request->getAll())) {
                     $this->flashBag->set('success', 'The element has been updated successfully');
+
+                    $historyService->write('News', 'Post "%s" has been updated', $name);
                     return '1';
                 }
 
             } else {
                 if ($service->add($this->request->getAll())) {
                     $this->flashBag->set('success', 'The element has been created successfully');
+
+                    $historyService->write('News', 'New post "%s" has been created', $name);
                     return $service->getLastId();
                 }
             }